pyproject.toml
[project]
name = "saidnedushi"
version = "0.1.0"
description = ""
authors = [
    {name = "Клим Клушин",email = "klim.klushin@gmail.com"}
]
#readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all] (>=0.116.1,<0.117.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
